Tutorial 4 - Create Results Form
Also added URL in mysite to land us on index.html for any incoming url with nothing specified.  I don't know if I missed this in an earlier tutorial or if it was left blank by the authors.  I DO know that it bugs me not to have a default page - so now it is the Polls index.html :-)